[FEAT] TaskCategoryRestController POST, PUT and DELETE method for CRUD

+++ src/main/java/it/mytools/spring/db/serv/TaskCategoryService.java
@@ -23,10 +23,16 @@ public class TaskCategoryService {
     }
 
     public void save(TaskCategory taskCategory) {
+        if (taskCategory == null) {
+            return;
+        }
         taskCategoryRepository.save(taskCategory);
     }
 
     public void delete(TaskCategory taskCategory) {
+        if (taskCategory == null) {
+            return;
+        }
         taskCategoryRepository.delete(taskCategory);
     }
 }
